# src/main/java/com/animesh/employee/service/resource/Role.java
package com.animesh.employee.service.resource;

import jakarta.validation.constraints.NotEmpty;
import lombok.Data;

@Data
public class Role {
    private String id;
    @NotEmpty
    private String name;
}